requirements.txt
flask==3.0.3
pika==1.3.2
pandas==2.2.2
google-cloud-storage==2.18.2
schema==0.7.7
psycopg2-binary==2.9.9
python-dotenv
xlwt
xlsxwriter